fix: 하단 모달과 댓글 작성 컴포넌트 겹치는 문제 fix
