+++ src/components/post/CommentAddBox.jsx
@@ -1,24 +1,25 @@
 import React, { useRef } from 'react';
 import styled from 'styled-components';
-import basicProfile from '../../assets/basic-profile.png';
+import UserProfile from '../profileImg/UserProfileImg';
 
 const Wrap = styled.div`
-  width: 100%;
+  width: 390px;
   position: fixed;
   bottom: 0;
+  padding: 12.5px 20px 12px 16px;
   border-top: 0.5px solid var(--border-gray);
   background-color: white;
 `;
 
-const AuthorProfile = styled.img`
+const AuthorProfile = styled(UserProfile)`
   width: 36px;
-  margin: 13px 18px 12px 16px;
+  margin-right: 16px;
   vertical-align: top;
 `;
 
 const CommentInput = styled.input`
   display: inline-block;
-  padding: 23px 0;
+  margin-top: 10.5px;
   font-family: 'Spoqa Han Sans Neo';
   font-weight: 400;
   font-size: 14px;
@@ -36,7 +37,7 @@ const CommentInput = styled.input`
 `;
 
 const AddBtn = styled.button`
-  padding: 23px 16px 20px;
+  margin-top: 10.5px;
   font-family: 'Spoqa Han Sans Neo';
   font-weight: 500;
   font-size: 14px;
@@ -45,16 +46,23 @@ const AddBtn = styled.button`
   float: right;
 `;
 
-export default function CommentAddBox() {
+export default function CommentAddBox({ ...props }) {
   const inputRef = useRef();
-  //   const [addBtnAble, setAddBtnAble] = useState(true);
+
+  const userInfo = JSON.parse(localStorage.getItem('userInfo'));
+  const { image } = userInfo.user;
+  if (!userInfo) {
+    console.log('로그인을 먼저 해주세요.');
+    props.history.push('/login');
+    return;
+  }
 
   return (
     <Wrap>
-      <AuthorProfile src={basicProfile} alt='작성자 프로필 이미지' />
+      <AuthorProfile size='tiny' src={image} alt='댓글 작성자 프로필 이미지' />
       <CommentInput
         placeholder='댓글 입력하기'
-        size='75'
+        size='35'
         useRef={inputRef}
         required='required'
       />
