+++ src/components/post/CommentItem.jsx
@@ -1,9 +1,7 @@
-import React, { useState } from 'react';
+import React from 'react';
 import styled, { css } from 'styled-components';
 import basicProfile from '../../assets/basic-profile.png';
 import viewMore from '../../assets/icon/icon-more-vertical.svg';
-import ModalContainer from '../common/modal/ModalContainer';
-import ModalList from '../common/modal/ModalList';
 
 const FontFamily = css`
   font-family: 'Spoqa Han Sans Neo';
@@ -57,11 +55,7 @@ const Comment = styled.p`
   word-break: break-all;
 `;
 
-export default function CommentItem() {
-  const [modalOpen, setModalOpen] = useState(false);
-  const handleModalOpen = () => {
-    setModalOpen(!modalOpen);
-  };
+export default function CommentItem({ onClick }) {
   return (
     <>
       <CommentItemWrap>
@@ -69,7 +63,7 @@ export default function CommentItem() {
           <UserProfile src={basicProfile} alt='댓글 작성자 프로필 이미지' />
           <UserName>서귀포시 무슨 농장</UserName>
           <CreatedTime>5분 전</CreatedTime>
-          <ViewMoreBtn onClick={handleModalOpen}>
+          <ViewMoreBtn onClick={onClick}>
             <ViewMore src={viewMore} alt='더 보기' />
           </ViewMoreBtn>
         </UserWrap>
@@ -79,11 +73,6 @@ export default function CommentItem() {
           수 있나요? 기다리기 지쳤어요 땡뻘땡뻘...
         </Comment>
       </CommentItemWrap>
-      {modalOpen && (
-        <ModalContainer>
-          <ModalList>신고하기</ModalList>
-        </ModalContainer>
-      )}
     </>
   );
 }
