+++ src/pages/PostPage.jsx
@@ -1,14 +1,17 @@
-import React from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
 import TopNavBasic from '../components/common/nav/TopNavBasic';
 import PostItem from '../components/post/PostItem';
 import CommentAddBox from '../components/post/CommentAddBox';
 import CommentItem from '../components/post/CommentItem';
+import ModalContainer from '../components/common/modal/ModalContainer';
+import ModalList from '../components/common/modal/ModalList';
 
 const PostItemWrap = styled.main`
   margin: 20px 16px 24px;
 `;
 const CommentListWrap = styled.section`
+  position: relative;
   margin: 20px 16px 66px;
   &::before {
     content: '';
@@ -18,13 +21,23 @@ const CommentListWrap = styled.section`
     background-color: var(--border-gray);
   }
 `;
+
 export default function PostPage({ location, ...props }) {
+  const [modalOpen, setModalOpen] = useState(false);
+  const handleModalOpen = () => {
+    setModalOpen(!modalOpen);
+  };
+
   return (
     <>
       <TopNavBasic viewMore />
       <PostItemWrap>
         <PostItem post={location.state.post} />
         <CommentListWrap>
+          <CommentItem onClick={handleModalOpen} />
+          <CommentItem />
+          <CommentItem />
+          <CommentItem />
           <CommentItem />
           <CommentItem />
           <CommentItem />
@@ -33,6 +46,11 @@ export default function PostPage({ location, ...props }) {
         </CommentListWrap>
       </PostItemWrap>
       <CommentAddBox color={'var(--border-gray)'} />
+      {modalOpen && (
+        <ModalContainer>
+          <ModalList>신고하기</ModalList>
+        </ModalContainer>
+      )}
     </>
   );
 }
